package/ltp-testsuite: simplify code to exclude test-cases
We have three conditions under which some tests may get removed:
uClibc, musl, static libraries. All three use the same mechanism
to exclude those test-cases: remove the files.

The first two use a common variable to list the affected files,
and share the same hook of their own to iterate over that list,
while the third has its own hook.

This is not very clean, so switch to using a single variable and
a single hook for all three conditions.

Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
Cc: Petr Vorel <petr.vorel@gmail.com>
Reviewed-by: Petr Vorel <petr.vorel@gmail.com>