package/ply: fix dependencies of comment for dependencies
Commits c86029ad7d9e (package/ply: needs headers >= 4.14) and
83d9fa314c46 (package/ply: needs dynamic library) added restrictions
on the availability of ply. The first forgot to add a comment, and
the second mis-handled the dependency on the headers version.

Indeed, we want the comment to show the requirement on the headers
version (since that is not a hardware dependency).

Fix this comment to include the headers version, and fix the condition
accordingly.

Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
Cc: Andreas Klinger <ak@it-klinger.de>